+++ api/src/controllers/index.ts
@@ -1,5 +1,11 @@
 import placesController from './places';
+import filtersController from './filters';
+import usersController from './users';
+import authController from './auth';
 
 export {
     placesController,
+    filtersController,
+    usersController,
+    authController,
 };
