+++ api/src/routes/filters.ts
@@ -1,11 +1,7 @@
 import express from 'express';
 
 import { authorized } from '../interceptors';
-import { translateText } from '../util';
-
-import {
-    Categories, CompanySizes, RestCosts, RestDurations,
-} from '../staticModels';
+import { filtersController } from '../controllers';
 
 const router = express.Router();
 
@@ -23,25 +19,6 @@ const router = express.Router();
  *        - default: []
  *
  */
-router.get('/', authorized, (request, response) => {
-    const categories = Categories.getAll()
-        .map((category) => ({ id: category.id, name: translateText(category.nameTextId, request.locale) }));
-
-    const costs = RestCosts.getAll()
-        .map((cost) => ({ id: cost.id, name: translateText(cost.nameTextId, request.locale) }));
-
-    const restDurations = RestDurations.getAll()
-        .map((duration) => ({ id: duration.id, name: translateText(duration.nameTextId, request.locale) }));
-
-    const companySizes = CompanySizes.getAll()
-        .map((companySize) => ({ id: companySize.id, name: translateText(companySize.nameTextId, request.locale) }));
-
-    response.json({
-        categories,
-        costs,
-        restDurations,
-        companySizes,
-    });
-});
+router.get('/', authorized, filtersController.getFilters);
 
 export default router;
